Test downloading of sequences in a given bbox
The new testcase starts the MapillarySequenceDownloadThread, waits until it's finished. Then it checks, if images were added to the MapillaryLayer. We can assume, that there are images in the given rectangle, as currently there are several images uploaded there.
This change required quite extensive changes, as the tests are run in headless mode and so the GUI must not be started. TestUtil is built along the lines of JosmFixture from josm-core, but trimmed a lot.

git-svn-id: http://svn.openstreetmap.org/applications/editors/josm/plugins@31374 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4